requirements.txt
django==3.2
djangorestframework==3.14.0
nltk==3.8.1